+++ example/javascript/router-zod.js
@@ -1,10 +1,12 @@
-'use strict'
+import express from 'express';
+import users from './users.js';
+import { z } from 'zod';
+import _ from 'lodash';
+import bodyParser from 'body-parser';
+import { createValidator } from '../../express-standard-schema-validation.js';
 
-const route = (module.exports = require('express').Router())
-const users = require('./users')
-const { z } = require('zod')
-const _ = require('lodash')
-const validator = require('../../').createValidator()
+const route = express.Router();
+const validator = createValidator();
 
 /**
  * This "GET /:id" endpoint is used to query users by their ID using Zod
@@ -12,27 +14,25 @@ const validator = require('../../').createValidator()
  * Now try http://localhost:8080/users/bananas - it will fail since the ID must be a number
  */
 const paramsSchema = z.object({
-  id: z.coerce.number()
-})
+  id: z.coerce.number(),
+});
 
 route.get('/:id', validator.params(paramsSchema), (req, res) => {
-  console.log(`\nGetting user by ID ${req.params.id}.`)
-  console.log(
-    `req.params was ${JSON.stringify(req.originalParams)} before validation`
-  )
-  console.log(`req.params is ${JSON.stringify(req.params)} after validation`)
-  console.log('note that the ID was correctly coerced to a number')
+  console.log(`\nGetting user by ID ${req.params.id}.`);
+  console.log(`req.params was ${JSON.stringify(req.originalParams)} before validation`);
+  console.log(`req.params is ${JSON.stringify(req.params)} after validation`);
+  console.log('note that the ID was correctly coerced to a number');
 
-  const u = _.find(users, { id: req.params.id })
+  const u = _.find(users, { id: req.params.id });
 
   if (u) {
-    res.json(u)
+    res.json(u);
   } else {
     res.status(404).json({
-      message: `no user exists with id "${req.params.id}"`
-    })
+      message: `no user exists with id "${req.params.id}"`,
+    });
   }
-})
+});
 
 /**
  * This "GET /" endpoint is used to query users by a querystring using Zod
@@ -40,37 +40,22 @@ route.get('/:id', validator.params(paramsSchema), (req, res) => {
  * Now try http://localhost:8080/users - it will fail since name is required
  */
 const querySchema = z.object({
-  name: z
-    .string()
-    .min(1)
-    .max(10),
-  age: z.coerce
-    .number()
-    .int()
-    .min(1)
-    .max(120)
-    .optional()
-})
+  name: z.string().min(1).max(10),
+  age: z.coerce.number().int().min(1).max(120).optional(),
+});
 
 route.get('/', validator.query(querySchema), (req, res) => {
-  console.log(`\nGetting users for query ${JSON.stringify(req.query)}.`)
-  console.log(
-    `req.query was ${JSON.stringify(req.originalQuery)} before validation`
-  )
-  console.log(`req.query is ${JSON.stringify(req.query)} after validation`)
-  console.log(
-    'note that the age was correctly coerced to a number if provided\n'
-  )
+  console.log(`\nGetting users for query ${JSON.stringify(req.query)}.`);
+  console.log(`req.query was ${JSON.stringify(req.originalQuery)} before validation`);
+  console.log(`req.query is ${JSON.stringify(req.query)} after validation`);
+  console.log('note that the age was correctly coerced to a number if provided\n');
 
   res.json(
-    _.filter(users, u => {
-      return (
-        _.includes(u.name, req.query.name) ||
-        (req.query.age && u.age === req.query.age)
-      )
-    })
-  )
-})
+    _.filter(users, (u) => {
+      return _.includes(u.name, req.query.name) || (req.query.age && u.age === req.query.age);
+    }),
+  );
+});
 
 /**
  * This "POST /" endpoint is used to create new users using Zod
@@ -78,37 +63,23 @@ route.get('/', validator.query(querySchema), (req, res) => {
  * Now try posting '{"name": "jane", "age": 1000}' - it will fail since the age is above 120
  */
 const bodySchema = z.object({
-  name: z
-    .string()
-    .min(1)
-    .max(10),
-  age: z.coerce
-    .number()
-    .int()
-    .min(1)
-    .max(120)
-})
+  name: z.string().min(1).max(10),
+  age: z.coerce.number().int().min(1).max(120),
+});
 
-route.post(
-  '/',
-  require('body-parser').json(),
-  validator.body(bodySchema),
-  (req, res) => {
-    console.log(`\nCreating user with data ${JSON.stringify(req.body)}.`)
-    console.log(
-      `req.body was ${JSON.stringify(req.originalBody)} before validation`
-    )
-    console.log(`req.body is ${JSON.stringify(req.body)} after validation`)
-    console.log(
-      'note that the age was correctly coerced to a number if it was a string\n'
-    )
+route.post('/', bodyParser.json(), validator.body(bodySchema), (req, res) => {
+  console.log(`\nCreating user with data ${JSON.stringify(req.body)}.`);
+  console.log(`req.body was ${JSON.stringify(req.originalBody)} before validation`);
+  console.log(`req.body is ${JSON.stringify(req.body)} after validation`);
+  console.log('note that the age was correctly coerced to a number if it was a string\n');
 
-    // Generate data required for insert (new id is incremented from previous max)
-    const prevMaxId = _.maxBy(users, u => u.id).id
-    const data = Object.assign({}, req.body, { id: prevMaxId + 1 })
+  // Generate data required for insert (new id is incremented from previous max)
+  const prevMaxId = _.maxBy(users, (u) => u.id).id;
+  const data = Object.assign({}, req.body, { id: prevMaxId + 1 });
 
-    users.push(data)
+  users.push(data);
 
-    res.json({ message: 'created user', data: data })
-  }
-)
+  res.json({ message: 'created user', data: data });
+});
+
+export default route;