+++ example/javascript/users.js
@@ -1,6 +1,4 @@
-'use strict'
-
-module.exports = [
+const users = [
   { id: 1000, name: 'anne, a.', age: 25 },
   { id: 1001, name: 'barry, a.', age: 52 },
   { id: 1002, name: 'clare, a.', age: 25 },
@@ -8,5 +6,7 @@ module.exports = [
   { id: 1004, name: 'anne, b.', age: 47 },
   { id: 1005, name: 'barry, b.', age: 80 },
   { id: 1006, name: 'clare, b.', age: 28 },
-  { id: 1007, name: 'joe, b.', age: 15 }
-]
+  { id: 1007, name: 'joe, b.', age: 15 },
+];
+
+export default users;