+++ prettier.config.js
@@ -1,0 +1,12 @@
+const config = {
+  arrowParens: 'always',
+  bracketSpacing: true,
+  jsxSingleQuote: true,
+  printWidth: 120,
+  semi: true,
+  singleQuote: true,
+  tabWidth: 2,
+  useTabs: false,
+};
+
+export default config;